added settings page + update button

+++ src/App/Updater.ts
@@ -1,12 +1,15 @@
 import {getLatestRelease} from "../Utils/HRemoteProfiles";
 import {downloadAsync} from "../Utils/HDownload";
-import {app} from "electron";
+import {app} from "@electron/remote";
 import path from "path";
 import cp from "child_process"
 import fs from "fs/promises";
 const window = require("../Interface/UIElements/scripts/window.js")
 
 let githubReleaseData: any = null;
+export let updateAvailable = false;
+export let newVersion: string
+
 
 export async function checkForUpdate() {
     return new Promise<boolean>(async (resolve, reject) => {
@@ -14,33 +17,35 @@ export async function checkForUpdate() {
             const currentVersion = require("../../package.json").version;
             const latestVersion = res["tag_name"];
 
-            if(currentVersion !== latestVersion) {
-                const settings = document.getElementById("settings")
-                settings!.setAttribute("badge", "");
+            githubReleaseData = res
 
-                resolve(true)
-            }
+            updateAvailable = currentVersion !== latestVersion;
 
-            resolve(false)
+            newVersion = latestVersion
+            resolve(updateAvailable)
         }).catch((err) => reject(err))
     })
 }
 
 export async function updateCli() {
     return new Promise<void>(async (resolve, reject) => {
-        const dlUrl = githubReleaseData["assets"][0]["browser_download_url"]
-        const name = githubReleaseData["assets"][0]["name"]
+        if(githubReleaseData) {
+            const dlUrl = githubReleaseData["assets"][0]["browser_download_url"]
+            const name = githubReleaseData["assets"][0]["name"]
 
-        await downloadAsync(dlUrl, path.join(app.getPath("temp"), name)).then((installerPath) => {
-            const child = cp.exec(`${installerPath} /S /LAUNCH`)
+            await downloadAsync(dlUrl, path.join(app.getPath("temp"), name)).then((installerPath) => {
+                const child = cp.exec(`${installerPath} /S /LAUNCH`)
 
-            child.on("error", (err) => {
-                reject(err)
-            })
+                child.on("error", (err) => {
+                    reject(err)
+                })
 
-            child.on("exit", async () => {
-                await fs.rm(installerPath).finally(() => app.quit())
-            })
-        }).catch((err) => reject(err))
+                child.on("exit", async () => {
+                    await fs.rm(installerPath).finally(() => app.quit())
+                })
+            }).catch((err) => reject(err))
+        } else {
+            console.error(`Impossible de mettre à jour le client, effectuez une vérification de mise à jour avant d'en lancer une`)
+        }
     })
 }
